apps/client_backend/requirements.txt
vocode-0.1.110-py3-none-any.whl
python-dotenv
redis